added git-config.sh and added checks in create_sym_links.sh to check if file exists

+++ bin/create_sym_links.sh
@@ -1,1 +1,8 @@
-ln -s ~/ubuntu_setup/env/dot_inputrc ~/.inputrc
+echo -n "Creating sym link ~/.inputrc to ~/ubuntu_setup/env/dot_inputrc..."
+if [ -f /home/andrewc1204/.inputrc ];
+then
+  echo "file already exists."
+else
+  ln -s ~/ubuntu_setup/env/dot_inputrc ~/.inputrc
+  echo "done."
+fi
